+++ kola/tests/docker/docker.go
@@ -66,8 +66,7 @@ func init() {
 		ClusterSize: 2,
 		Name:        "docker.network",
 		Distros:     []string{"cl"},
-
-		// qemu-unpriv machines cannot communicate
+		// No idea why Docker containers cannot reach each the other VM
 		ExcludePlatforms: []string{"qemu-unpriv"},
 	})
 	register.Register(&register.Test{
@@ -109,9 +108,6 @@ storage:
 passwd:
   users:
   - name: dockremap`),
-
-		// qemu-unpriv machines cannot communicate
-		ExcludePlatforms: []string{"qemu-unpriv"},
 	})
 
 	// This test covers all functionality that should be quick to run and can be
@@ -223,12 +219,6 @@ systemd:
   units:
    - name: docker.service
      enable: true`),
-
-		// https://github.com/coreos/mantle/issues/999
-		// On the qemu-unpriv platform the DHCP provides no data, pre-systemd 241 the DHCP server sending
-		// no routes to the link to spin in the configuring state. docker.service pulls in the network-online
-		// target which causes the basic machine checks to fail
-		ExcludePlatforms: []string{"qemu-unpriv"},
 	})
 }