+++ kola/tests/systemd/journald.go
@@ -62,8 +62,7 @@ func init() {
 
 		// Disabled on Azure because setting hostname
 		// is required at the instance creation level
-		// qemu-unpriv machines cannot communicate
-		ExcludePlatforms: []string{"azure", "qemu-unpriv"},
+		ExcludePlatforms: []string{"azure"},
 	})
 }